ai-engine/requirements.txt
kafka-python
pandas==1.5.3
scikit-learn==1.2.2
joblib==1.2.0
requests==2.31.0
